pyproject.toml
[tool.poetry]
name = "companiesvacanciesdatabase"
version = "0.1.0"
description = ""
authors = ["Anastasiya Soloviova <mafanya.as@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
mypy = "^1.15.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 119

[tool.black]
line-length = 119
exclude = ''' ( /( \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist )/
    | foo.py
      ) '''

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
